move statsd config validation to statsd.New
