+++ internal/statsd/main_test.go
@@ -6,10 +6,13 @@
 package statsd
 
 import (
+	"errors"
+	"strings"
 	"testing"
 	"time"
 
 	"github.com/circonus-labs/circonus-agent/internal/config"
+	"github.com/circonus-labs/circonus-agent/internal/config/defaults"
 	"github.com/rs/zerolog"
 	"github.com/spf13/viper"
 )
@@ -32,16 +35,45 @@ func TestNew(t *testing.T) {
 		viper.Reset()
 	}
 
-	t.Log("Enabled")
+	t.Log("Enabled - no port")
 	{
 		viper.Set(config.KeyStatsdDisabled, false)
+		expect := errors.New("Invalid StatsD port (empty)")
+		_, err := New()
+		if err == nil {
+			t.Fatal("expect error")
+		}
+		if err.Error() != expect.Error() {
+			t.Fatalf("expected (%s) got (%s)", expect, err)
+		}
+		viper.Reset()
+	}
+
+	t.Log("Enabled - default port, invalid host category")
+	{
+		viper.Set(config.KeyStatsdDisabled, false)
+		viper.Set(config.KeyStatsdPort, defaults.StatsdPort)
+		expect := errors.New("Invalid StatsD host category (empty)")
+		_, err := New()
+		if err == nil {
+			t.Fatal("expect error")
+		}
+		if err.Error() != expect.Error() {
+			t.Fatalf("expected (%s) got (%s)", expect, err)
+		}
+		viper.Reset()
+	}
+
+	t.Log("Enabled - default port, default host category")
+	{
+		viper.Set(config.KeyStatsdDisabled, false)
+		viper.Set(config.KeyStatsdPort, defaults.StatsdPort)
+		viper.Set(config.KeyStatsdHostCategory, defaults.StatsdHostCategory)
 		s, err := New()
 		if err != nil {
 			t.Fatalf("expected NO error, got (%s)", err)
 		}
-		if s == nil {
-			t.Fatal("expected not nil")
-		}
+		s.listener.Close()
 		viper.Reset()
 	}
 }
@@ -65,6 +97,8 @@ func TestStart(t *testing.T) {
 	t.Log("Enabled w/Stop")
 	{
 		viper.Set(config.KeyStatsdDisabled, false)
+		viper.Set(config.KeyStatsdPort, defaults.StatsdPort)
+		viper.Set(config.KeyStatsdHostCategory, defaults.StatsdHostCategory)
 		s, err := New()
 		if err != nil {
 			t.Fatalf("expected NO error, got (%s)", err)
@@ -72,7 +106,7 @@ func TestStart(t *testing.T) {
 		if s == nil {
 			t.Fatal("expected not nil")
 		}
-		time.AfterFunc(2*time.Second, func() {
+		time.AfterFunc(1*time.Second, func() {
 			s.Stop()
 		})
 		s.Start()
@@ -82,6 +116,8 @@ func TestStart(t *testing.T) {
 	t.Log("Enabled w/direct server close")
 	{
 		viper.Set(config.KeyStatsdDisabled, false)
+		viper.Set(config.KeyStatsdPort, defaults.StatsdPort)
+		viper.Set(config.KeyStatsdHostCategory, defaults.StatsdHostCategory)
 		s, err := New()
 		if err != nil {
 			t.Fatalf("expected NO error, got (%s)", err)
@@ -89,7 +125,7 @@ func TestStart(t *testing.T) {
 		if s == nil {
 			t.Fatal("expected not nil")
 		}
-		time.AfterFunc(2*time.Second, func() {
+		time.AfterFunc(1*time.Second, func() {
 			s.listener.Close()
 		})
 		s.Start()
@@ -116,6 +152,8 @@ func TestStop(t *testing.T) {
 	t.Log("Enabled")
 	{
 		viper.Set(config.KeyStatsdDisabled, false)
+		viper.Set(config.KeyStatsdPort, defaults.StatsdPort)
+		viper.Set(config.KeyStatsdHostCategory, defaults.StatsdHostCategory)
 		s, err := New()
 		if err != nil {
 			t.Fatalf("expected NO error, got (%s)", err)
@@ -123,7 +161,7 @@ func TestStop(t *testing.T) {
 		if s == nil {
 			t.Fatal("expected not nil")
 		}
-		time.AfterFunc(2*time.Second, func() {
+		time.AfterFunc(1*time.Second, func() {
 			s.Stop()
 		})
 		s.Start()
@@ -154,31 +192,35 @@ func TestFlush(t *testing.T) {
 	t.Log("Flush (no stats)")
 	{
 		viper.Set(config.KeyStatsdDisabled, false)
+		viper.Set(config.KeyStatsdPort, defaults.StatsdPort)
+		viper.Set(config.KeyStatsdHostCategory, defaults.StatsdHostCategory)
 		s, err := New()
 		if err != nil {
 			t.Fatalf("expected NO error, got (%s)", err)
 		}
 		metrics := s.Flush()
-		viper.Reset()
-
+		s.listener.Close()
 		if metrics == nil {
 			t.Fatal("expected not nil")
 		}
 		if len(*metrics) != 0 {
 			t.Fatalf("expected empty metrics, got (%#v)", metrics)
 		}
+		viper.Reset()
 	}
 
 	t.Log("Flush (no stats, nil hostMetrics)")
 	{
 		viper.Set(config.KeyStatsdDisabled, false)
+		viper.Set(config.KeyStatsdPort, defaults.StatsdPort)
+		viper.Set(config.KeyStatsdHostCategory, defaults.StatsdHostCategory)
 		s, err := New()
 		if err != nil {
 			t.Fatalf("expected NO error, got (%s)", err)
 		}
 		s.hostMetrics = nil
 		metrics := s.Flush()
-		viper.Reset()
+		s.listener.Close()
 
 		if metrics == nil {
 			t.Fatal("expected not nil")
@@ -186,5 +228,339 @@ func TestFlush(t *testing.T) {
 		if len(*metrics) != 0 {
 			t.Fatalf("expected empty metrics, got (%#v)", metrics)
 		}
+		viper.Reset()
 	}
+}
+
+func TestValidateStatsdOptions(t *testing.T) {
+	t.Log("Testing validateStatsdOptions")
+
+	viper.Set(config.KeyStatsdDisabled, true)
+
+	t.Log("StatsD disabled")
+	{
+		err := validateStatsdOptions()
+		if err != nil {
+			t.Fatalf("Expected NO error, got (%v)", err)
+		}
+	}
+
+	viper.Set(config.KeyStatsdDisabled, false)
+
+	t.Log("StatsD port (invalid, empty)")
+	{
+		viper.Set(config.KeyStatsdPort, "")
+
+		expectedErr := errors.New("Invalid StatsD port (empty)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("StatsD port (invalid, not a number)")
+	{
+		viper.Set(config.KeyStatsdPort, "abc")
+
+		expectedErr := errors.New("Invalid StatsD port (abc)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("StatsD port (invalid, out of range, low)")
+	{
+		viper.Set(config.KeyStatsdPort, "10")
+
+		expectedErr := errors.New("Invalid StatsD port 1024>10<65535")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("StatsD port (invalid, out of range, high)")
+	{
+		viper.Set(config.KeyStatsdPort, "70000")
+
+		expectedErr := errors.New("Invalid StatsD port 1024>70000<65535")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	viper.Set(config.KeyStatsdPort, "8125")
+
+	t.Log("Host category (invalid, empty)")
+	{
+		viper.Set(config.KeyStatsdHostCategory, "")
+
+		expectedErr := errors.New("Invalid StatsD host category (empty)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	viper.Set(config.KeyStatsdHostCategory, "statsd")
+
+	t.Log("Group CID, OK - none")
+	{
+		viper.Set(config.KeyStatsdGroupCID, "")
+		err := validateStatsdOptions()
+		if err != nil {
+			t.Fatalf("Expected NO error, got (%v)", err)
+		}
+	}
+
+	t.Log("Group CID (cosi, no cfg)")
+	{
+		viper.Set(config.KeyStatsdGroupCID, "cosi")
+
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if !strings.HasPrefix(err.Error(), "Unable to access cosi check config:") {
+			t.Errorf("Expected (%s) got (%s)", "Unable to access cosi check config: ...", err)
+		}
+	}
+
+	t.Log("Group CID (invalid, abc)")
+	{
+		viper.Set(config.KeyStatsdGroupCID, "abc")
+
+		expectedErr := errors.New("Invalid StatsD Group Check ID (abc)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("Group CID, valid - 123")
+	{
+		viper.Set(config.KeyStatsdGroupCID, "123")
+		expectedErr := errors.New("StatsD host/group prefix mismatch (both empty)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("Group CID, valid - /check_bundle/123")
+	{
+		viper.Set(config.KeyStatsdGroupCID, "/check_bundle/123")
+		expectedErr := errors.New("StatsD host/group prefix mismatch (both empty)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	viper.Set(config.KeyStatsdGroupCID, "/check_bundle/123")
+	viper.Set(config.KeyStatsdHostPrefix, "host.")
+
+	t.Log("Group prefix, invalid (same as host)")
+	{
+		viper.Set(config.KeyStatsdGroupPrefix, "host.")
+
+		expectedErr := errors.New("StatsD host/group prefix mismatch (same)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	viper.Set(config.KeyStatsdGroupPrefix, "group.")
+
+	t.Log("Group counter operator, invalid (empty)")
+	{
+		viper.Set(config.KeyStatsdGroupCounters, "")
+
+		expectedErr := errors.New("Invalid StatsD counter operator (empty)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("Group counter operator, invalid ('multiply')")
+	{
+		viper.Set(config.KeyStatsdGroupCounters, "multiply")
+
+		expectedErr := errors.New("Invalid StatsD counter operator (multiply)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("Group counter operator, invalid ('sum')")
+	{
+		viper.Set(config.KeyStatsdGroupCounters, "sum")
+
+		expectedErr := errors.New("Invalid StatsD gauge operator (empty)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("Group counter operator, invalid ('average')")
+	{
+		viper.Set(config.KeyStatsdGroupCounters, "average")
+
+		expectedErr := errors.New("Invalid StatsD gauge operator (empty)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("Group gauge operator, invalid (empty)")
+	{
+		viper.Set(config.KeyStatsdGroupGauges, "")
+
+		expectedErr := errors.New("Invalid StatsD gauge operator (empty)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("Group gauge operator, invalid ('multiply')")
+	{
+		viper.Set(config.KeyStatsdGroupGauges, "multiply")
+
+		expectedErr := errors.New("Invalid StatsD gauge operator (multiply)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("Group gauge operator, invalid ('sum')")
+	{
+		viper.Set(config.KeyStatsdGroupGauges, "sum")
+
+		expectedErr := errors.New("Invalid StatsD set operator (empty)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("Group gauge operator, invalid ('average')")
+	{
+		viper.Set(config.KeyStatsdGroupGauges, "average")
+
+		expectedErr := errors.New("Invalid StatsD set operator (empty)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("Group set operator, invalid (empty)")
+	{
+		viper.Set(config.KeyStatsdGroupSets, "")
+
+		expectedErr := errors.New("Invalid StatsD set operator (empty)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("Group set operator, invalid ('multiply')")
+	{
+		viper.Set(config.KeyStatsdGroupSets, "multiply")
+
+		expectedErr := errors.New("Invalid StatsD set operator (multiply)")
+		err := validateStatsdOptions()
+		if err == nil {
+			t.Fatal("Expected error")
+		}
+		if err.Error() != expectedErr.Error() {
+			t.Errorf("Expected (%s) got (%s)", expectedErr, err)
+		}
+	}
+
+	t.Log("Group gauge operator, invalid ('sum')")
+	{
+		viper.Set(config.KeyStatsdGroupSets, "sum")
+
+		err := validateStatsdOptions()
+		if err != nil {
+			t.Fatalf("Expected NO error, got (%v)", err)
+		}
+	}
+
+	t.Log("Group gauge operator, invalid ('average')")
+	{
+		viper.Set(config.KeyStatsdGroupSets, "average")
+
+		err := validateStatsdOptions()
+		if err != nil {
+			t.Fatalf("Expected NO error, got (%v)", err)
+		}
+	}
+
+	viper.Reset()
 }
